refactor: added concerto-graphics-core target
