+++ Src/Concerto/Graphics/Core/Defines.hpp
@@ -11,10 +11,10 @@
 #pragma warning(disable: 4251) // Disable warning about DLL interface needed
 #endif
 
-#ifdef CONCERTO_GRAPHICS_BUILD
-	#define CONCERTO_GRAPHICS_API CCT_EXPORT
+#ifdef CONCERTO_GRAPHICS_CORE_BUILD
+	#define CONCERTO_GRAPHICS_CORE_API CCT_EXPORT
 #else
-	#define CONCERTO_GRAPHICS_API CCT_IMPORT
-#endif // CONCERTO_GRAPHICS_BUILD
+	#define CONCERTO_GRAPHICS_CORE_API CCT_IMPORT
+#endif // CONCERTO_GRAPHICS_CORE_BUILD
 
 #endif //CONCERTO_GRAPHICS_DEFINES_HPP
