+++ Src/Concerto/Graphics/Core/PixelFormat.hpp
