+++ Src/Concerto/Graphics/Core/Version.hpp
@@ -5,7 +5,7 @@
 #ifndef CONCERTO_GRAPHICS_VERSION_HPP
 #define CONCERTO_GRAPHICS_VERSION_HPP
 
-#include "Concerto/Graphics/Defines.hpp"
+#include "Concerto/Graphics/Core/Defines.hpp"
 
 namespace cct::gfx
 {
