+++ Src/Concerto/Graphics/Core/Vertex.hpp
